\name{RobustSNP-package}
\alias{RobustSNP-package}
\alias{RobustSNP}
\docType{package}
\title{
Robust SNP association tests under different genetic models, allowing for covariates
}
\description{
This function performs robust genetic association tests. SNP associations are tested under 
additive, dominant and recessive models. Valid p-values are returned for each SNP, 
taking into account the multiple testing under different models. Binary or quantitative outcomes are allowed. 
Covariates may also be included. The program can be applied to genome-wide association studies (GWAS).    
}
\details{
\tabular{ll}{
Package: \tab RobustSNP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Depends: \tab mvtnorm,car,snpMatrix\cr
Date: \tab 2010-07-02\cr
License: \tab GPL(>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Hon-Cheong So
Maintainer: Hon-Cheong So <hcso85@gmail.com>

}
\references{
Hon-Cheong So and Pak C. Sham (2010).Robust association tests under different genetic models, allowing for 
binary or quantitative traits and covariates. Submitted. 
}

\keyword{ package }


